+++ R/twalk.R
@@ -468,7 +468,7 @@ GetAutoCov <- function( dt, lags) {
 
 #' @name IAT
 #' @title calculate the Integrated Autocorrelation Time
-#' @description Calculate the Tntegrated Autocorrelation Time, which gives the proposed value for thinning. E.g., if the IAT is 80, it is good to thin the MCMC run by storing only every 80 iterations. This method is slower than GetAutoCov, but much better.
+#' @description Calculate the Integrated Autocorrelation Time, which gives the proposed value for thinning. E.g., if the IAT is 80, it is good to thin the MCMC run by storing only every 80 iterations. This method is slower than GetAutoCov, but much better.
 #' @param set This option reads the 'info' variable, which contains the data and the model output.
 #' @param par The parameter to test. Defaults to 0.
 #' @param from The first of the iterations of the MCMC run to check. Defaults to the first one.